refactor: consolidate alias validation into config package
Consolidate alias-related code into the config package by:

1. File organization:
   - Renamed alias.go -> alias_parse.go (placeholder parsing & alias access)
   - Created alias_validate.go (all validation logic)
   - Updated validation.go (removed alias validation, kept other validations)

2. Moved internal/aliasvalidator -> internal/config:
   - commandValidator struct (was Validator)
   - Security checks for shell metacharacters
   - Command whitelist validation
   - Now internal to config package

3. Benefits:
   - Single responsibility: parsing vs validation clearly separated
   - Alias code consolidated in one package (was spread across 3 locations)
   - Reduced package dependencies (config no longer imports aliasvalidator)
   - Better encapsulation (validator is now package-private)

4. File sizes after refactor:
   - alias_parse.go: ~150 lines (placeholder & parsing logic)
   - alias_validate.go: ~155 lines (all validation logic)
   - validation.go: ~100 lines (config-wide validation orchestration)

All tests passing with comprehensive coverage for validation logic.
